requirements.txt
numpy
pandas
flask
requests
python-dotenv

